ajeitações
