--- init_manual.py
@@ -1,6 +1,0 @@
-from iniciar import init_db
-from app import app
-
-init_db(app)
-
-# rode este arquivo para inicializar o banco de dados manualmente
