+++ requirements.txt
@@ -1,34 +1,18 @@
-asarPy==1.0.1
+black==25.1.0
 blinker==1.9.0
 click==8.2.1
-Cython==3.1.1
-blinker==1.9.0
-click==8.2.1
+colorama==0.4.6
 flake8==7.3.0
 Flask==3.1.1
 Flask-Login==0.6.3
-iniconfig==2.1.0
 itsdangerous==2.2.0
 Jinja2==3.1.6
-Mako==1.3.8
-Markdown==3.5.2
-MarkupSafe==3.0.2
-meson==1.5.2
-packaging==25.0
-pluggy==1.6.0
-pycairo==1.28.0
-PyGObject==3.50.1
-pytest==8.4.1
-setuptools==80.9.0
-six==1.17.0
-Werkzeug==3.1.3
-wheel==0.46.1
 MarkupSafe==3.0.2
 mccabe==0.7.0
+mypy_extensions==1.1.0
 packaging==25.0
-pluggy==1.6.0
+pathspec==0.12.1
+platformdirs==4.3.8
 pycodestyle==2.14.0
 pyflakes==3.4.0
-Pygments==2.19.2
-pytest==8.4.1
 Werkzeug==3.1.3
